Updated TCLogging to remove the hard log4j dependencies
 - Supports Log4j  by default if it is in the classpath, otherwise uses SLF4J
 - API break: had to remove TCLogger.setLevel() method and some other public methods on TCLogger
 - Provided implementations of TCLogger and TCLogging for both Log4j and SLF4J
 - Removed the JMX Appender (confirmed with @lorban)
 - Client Runtime only contains SLF4J api and ships with externalized log4j and slf4j-log4j
