+++ common/src/main/java/com/tc/logging/BaseMessageDecoratorTCLogger.java
@@ -95,11 +95,6 @@ public abstract class BaseMessageDecoratorTCLogger implements TCLogger {
   }
 
   @Override
-  public void setLevel(LogLevel level) {
-    logger.setLevel(level);
-  }
-
-  @Override
   public LogLevel getLevel() {
     return logger.getLevel();
   }
