+++ common/src/main/java/com/tc/logging/DelegatingTCLogger.java
@@ -1,0 +1,21 @@
+package com.tc.logging;
+
+import java.io.File;
+
+/**
+ * @author Mathieu Carbou
+ */
+interface DelegatingTCLogger {
+
+  void setLogDirectory(File theDirectory, int processType);
+
+  void disableLocking();
+
+  TCLogger newLogger(String name);
+
+  void closeFileAppender();
+
+  TCLogger getConsoleLogger();
+
+  TCLogger getOperatorEventLogger();
+}
