+++ common/src/test/java/com/tc/net/protocol/transport/ConnectionHealthCheckReverseCallbackTest.java
@@ -18,7 +18,6 @@
  */
 package com.tc.net.protocol.transport;
 
-import com.tc.logging.LogLevelImpl;
 import com.tc.logging.TCLogger;
 import com.tc.logging.TCLogging;
 import com.tc.net.TCSocketAddress;
@@ -53,10 +52,6 @@ import java.util.Collections;
 import java.util.concurrent.CountDownLatch;
 
 public class ConnectionHealthCheckReverseCallbackTest extends TCTestCase {
-
-  static {
-    TCLogging.getLogger(ConnectionHealthCheckerImpl.class).setLevel(LogLevelImpl.DEBUG);
-  }
 
   private final TCLogger            logger               = TCLogging.getLogger(getClass());
