+++ common/src/test/java/com/tc/util/concurrent/ThrottledTaskExecutorTest.java
@@ -18,8 +18,6 @@
  */
 package com.tc.util.concurrent;
 
-import com.tc.logging.LogLevelImpl;
-import com.tc.logging.TCLogging;
 import com.tc.test.TCTestCase;
 import com.tc.util.Assert;
 
@@ -32,7 +30,6 @@ public class ThrottledTaskExecutorTest extends TCTestCase {
   @Override
   protected void setUp() throws Exception {
     super.setUp();
-    TCLogging.getLogger(ThrottledTaskExecutor.class).setLevel(LogLevelImpl.DEBUG);
   }
 
   public void testLinearExecute() {
