+++ tc-messaging/src/main/java/com/tc/logging/TCLogger.java
@@ -48,8 +48,6 @@ public interface TCLogger {
 
   boolean isInfoEnabled();
 
-  void setLevel(LogLevel level);
-
   LogLevel getLevel();
 
   String getName();
